Replace API-Server healthcheck with Heartbeat check
